+++ mailscanner/rep_message_listing.php
@@ -52,11 +52,11 @@ if (isset($_GET['orderdir']) && !validateInput(deepSanitizeInput($_GET['orderdir
     exit(__('dievalidate99'));
 }
 
-$sql = "
+$sql = '
  SELECT
   id AS id2,
-  DATE_FORMAT(timestamp, '" . DATE_FORMAT . ' ' . TIME_FORMAT . "') AS datetime,
-  from_address,";
+  timestamp,
+  from_address,';
 if (defined('DISPLAY_IP') && DISPLAY_IP) {
     $sql .= 'clientip,';
 }