+++ mailscanner/rep_message_ops.php
@@ -60,11 +60,11 @@ if (QUARANTINE_USE_FLAG) {
 }
 
 // SQL query
-$sql = "
+$sql = '
  SELECT
   id AS id2,
-  DATE_FORMAT(timestamp, '" . DATE_FORMAT . ' ' . TIME_FORMAT . "') AS datetime,
-  from_address,";
+  timestamp,
+  from_address,';
 if (defined('DISPLAY_IP') && DISPLAY_IP) {
     $sql .= 'clientip,';
 }